postgres-chatbot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
asttokens #@ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm #@ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
dataclasses-json==0.6.6
debugpy #@ file:///croot/debugpy_1690905042057/work
decorator #@ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro==1.9.0
exceptiongroup #@ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing #@ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
#google-ai-generativelanguage==0.6.4
#google-api-core==2.19.0
#google-api-python-client==2.129.0
#google-auth==2.29.0
#google-auth-httplib2==0.2.0
#google-generativeai==0.5.4
#googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.7.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata #@ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
ipykernel #@ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython #@ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
jedi #@ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client #@ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core #@ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
#langchain-google-genai==1.0.4
langchain-openai
langchain-text-splitters==0.2.0
langsmith==0.1.60
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline #@ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio #@ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso #@ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect #@ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare #@ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs #@ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
prompt-toolkit #@ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
proto-plus==1.23.0
protobuf==4.25.3
psutil #@ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722404069/work
psycopg2-binary==2.9.9
ptyprocess #@ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval #@ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments #@ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
python-dateutil #@ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq #@ file:///croot/pyzmq_1705605076900/work
referencing==0.35.1
requests==2.32.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six #@ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data #@ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.34.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado #@ file:///home/conda/feedstock_root/build_artifacts/tornado_1708363103305/work
tqdm==4.66.4
traitlets #@ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing-inspect==0.9.0
typing_extensions #@ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth #@ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
yarl==1.9.4
zipp #@ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
